+++ ErsatzTV.Application/Playouts/PlayoutNameViewModel.cs
@@ -7,6 +7,10 @@ public record PlayoutNameViewModel(
     ProgramSchedulePlayoutType PlayoutType,
     string ChannelName,
     string ChannelNumber,
+    ChannelProgressMode ProgressMode,
     string ScheduleName,
     string ExternalJsonFile,
-    Option<TimeSpan> DailyRebuildTime);
+    TimeSpan? DbDailyRebuildTime)
+{
+    public Option<TimeSpan> DailyRebuildTime => Optional(DbDailyRebuildTime);
+}