+++ ErsatzTV.Application/Playouts/Queries/GetAllPlayoutsHandler.cs
@@ -25,9 +25,10 @@ public class GetAllPlayoutsHandler : IRequestHandler<GetAllPlayouts, List<Playou
                     p.ProgramSchedulePlayoutType,
                     p.Channel.Name,
                     p.Channel.Number,
+                    p.Channel.ProgressMode,
                     p.ProgramScheduleId == null ? string.Empty : p.ProgramSchedule.Name,
                     p.ExternalJsonFile,
-                    Optional(p.DailyRebuildTime)))
+                    p.DailyRebuildTime))
             .ToListAsync(cancellationToken);
     }
 }